PF/back-end/requirements.txt
django==4.1.3
Pillow==9.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
geopy==2.3.0
pandas==1.5.1
django-cors-headers==3.13.0
